Split battle finish class and broke the menu a bit more.
They will be fixed and/or revamped in the next commits.
